+++ src/mx_push_back_minways.c
@@ -1,6 +1,6 @@
 #include "../inc/pathfinder.h"
 
-void mx_push_back_minways(const char *file, t_minways **list, unsigned long **minwaym) {
+void mx_push_back_minways(const char *file, t_minways **list, unsigned int **minwaym) {
     t_minways *node = mx_create_minwaynode(file, minwaym);
     t_minways *pl = *list;
